Introduce the "cid" property for collectiona
You cannot get collections by their "cid", like you do this with models,
but the "cid" can be used to track the same collection instance too. It
can be usefult in applications, which cache collection instances, or in
applications, which gather and visualize collection events, like
Backbone.Debugger or Marionette.Agent.